tests/test_user_preferences.py
import unittest
from unittest.mock import AsyncMock, patch
from discord.ext import commands
from cogs.user_cog import UserCog

class TestUserPreferences(unittest.TestCase):
    def setUp(self):
        self.bot = commands.Bot(command_prefix='&')
        self.cog = UserCog(self.bot)
        self.ctx = AsyncMock()

    def test_set_user_preferences(self):
        # Replace with your test cases for the set_user_preferences command
        pass

    def test_get_user_preferences(self):
        # Replace with your test cases for the get_user_preferences command
        pass

if __name__ == '__main__':
    unittest.main()
