requirements.txt
discord.py==1.7.3          # Discord API library
requests==2.26.0           # HTTP requests library
python-dotenv==0.19.1      # Load environment variables from a .env file
